+++ src/remotes/remotefactory.ts
@@ -1,11 +1,9 @@
-import { ZesaruxExtRemote } from './zesarux/zesaruxextremote';
-import { RemoteBase } from './remotebase';
-//import {ZxNextUsbSerialRemote} from './zxnext/zxnextusbserialremote';
+import {RemoteBase} from './remotebase';
 import {ZSimRemote} from './zsimulator/zsimremote';
-//import {ZxNextSocketRemote} from './zxnext/zxnextsocketremote';
 import {CSpectRemote} from './dzrpbuffer/cspectremote';
 import {Utility} from '../misc/utility';
 import {ZxNextSocketRemote} from './dzrpbuffer/zxnextsocketremote';
+import {ZesaruxRemote} from './zesarux/zesaruxremote';
 
 
 
@@ -15,13 +13,12 @@ import {ZxNextSocketRemote} from './dzrpbuffer/zxnextsocketremote';
 export class RemoteFactory {
 	/**
 	 * Factory method to create an emulator.
-	 * @param remoteType 'zrcp', 'zxnext' or 'zsim'. For 'zrcp' always the ZesaruxExtEmulator is created.
-	 * It will fallback to Zesarux if no ZesaruxExt is connected.
+	 * @param remoteType 'zrcp', 'cspect', 'zxnext' or 'zsim'.
 	 */
 	public static createRemote(remoteType: string) {
 		switch (remoteType) {
 			case 'zrcp':	// ZEsarUX Remote Control Protocol
-				RemoteFactory.setRemote(new ZesaruxExtRemote());
+				RemoteFactory.setRemote(new ZesaruxRemote());
 				break;
 			case 'cspect':	// CSpect socket
 				RemoteFactory.setRemote(new CSpectRemote());
@@ -46,14 +43,14 @@ export class RemoteFactory {
 	 * Sets the emulator variable.
 	 */
 	protected static setRemote(emulator: RemoteBase) {
-		Remote=emulator;
+		Remote = emulator;
 	}
 
 	/**
 	 * Clears the emulator variable.
 	 */
 	public static removeRemote() {
-		Remote=undefined as any;
+		Remote = undefined as any;
 	}
 
 }